external/noisegraph/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "noisegraph"
version = "0.1.0"
description = "Standalone log shipper + baseline + behavioral graph noise-reduction engine for SOC workflows."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Shakeeb Salman"}]
keywords = ["security", "soc", "logging", "observability", "siem", "anomaly-detection"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Information Technology",
  "Intended Audience :: System Administrators",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Topic :: Security",
  "Topic :: System :: Logging",
]
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.23",
  "pydantic>=2.6",
  "python-dateutil>=2.9",
  "orjson>=3.9",
  "PyYAML>=6.0",
  "httpx>=0.26",
  "watchfiles>=0.21",
  "rich>=13.7",
  "typer>=0.12",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.4",
]

[project.scripts]
noisegraph = "noisegraph.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["noisegraph*"]

[tool.ruff]
line-length = 100
